Django/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==24.9.0
gunicorn==22.0.0
inflection==0.5.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.6.0
XlsxWriter==3.2.0
